// frontend/src/types/Aluno.ts
export interface Aluno {
  Adm_Id: number;
  RU: number;
  Nome: string;
  Nome_oficial: string;
  Cd_turma: number;
  Turma: string;
  Tipo_Servico: string;
  Vl_Titulo: number;
  Vl_Recebido: number;
  Vl_Base_Comissao: number;
  RDC: number;
  Vl_comissao: number;
  Dt_Liquidacao: string;
  Dt_Vencimento: string;
  Sinal: string;
  Titulo: number;
  Cd_local: number;
  Local_nome: string;
}